requirements.txt
Flask
Flask-Cors
Werkzeug
Pillow
python-docx
PyPDF2
pdf2image
waitress
